[NCL-5702]: Move TemporaryBuildsCleaner integration tests to the new package
