+++ integration-test/src/test/java/org/jboss/pnc/integration_new/endpoint/BuildTest.java
@@ -37,9 +37,9 @@ import org.jboss.pnc.dto.GroupConfiguration;
 import org.jboss.pnc.dto.requests.GroupBuildRequest;
 import org.jboss.pnc.enums.BuildStatus;
 import org.jboss.pnc.enums.RebuildMode;
-import org.jboss.pnc.integration.mock.RemoteBuildsCleanerMock;
 import org.jboss.pnc.integration.utils.ResponseUtils;
 import org.jboss.pnc.integration_new.setup.Deployments;
+import org.jboss.pnc.mock.RemoteBuildsCleanerMock;
 import org.jboss.pnc.rest.api.parameters.BuildParameters;
 import org.jboss.pnc.rest.api.parameters.BuildsFilterParameters;
 import org.jboss.pnc.rest.api.parameters.GroupBuildParameters;
